requirements.txt
requests
colorama
tqdm
python-whois
